Move json-schema validation logic inside the `XMLDiagnostics` class
