+++ src/diagnostics/base/XMLDiagnostics.ts
@@ -1,9 +1,13 @@
 // Library
 import * as vscode from 'vscode';
-import { XMLValidator } from 'fast-xml-parser';
+import { XMLValidator, XMLParser } from 'fast-xml-parser';
+import { Validator, Schema } from 'jsonschema';
 
 // Classes
 import { Diagnostics } from './Diagnostics';
+
+// Helpers
+import { xml } from '../../helpers';
 
 // ---------------
 // XML DIAGNOSTICS
@@ -21,8 +25,28 @@ export abstract class XMLDiagnostics extends Diagnostics {
      * @param context - The extension context ({@linkcode vscode.ExtensionContext})
      * @returns A new instance of the {@linkcode XMLDiagnostics} class
      */
-    constructor(protected name: string, context: vscode.ExtensionContext) {
+    constructor(
+        protected name: string,
+        context: vscode.ExtensionContext,
+        private schema: Schema,
+    ) {
         super(name, context);
+    }
+
+
+    /**
+     * Creates diagnostic problems for the given document.
+     * 
+     * @param document The text document to create diagnostic problems for. (see {@linkcode vscode.TextDocument})
+     * @returns An array of diagnostic problems. (see {@linkcode vscode.Diagnostic})
+     */
+    protected override createProblems(document: vscode.TextDocument): vscode.Diagnostic[] {
+        // Validate the document using the XML Validator from `fast-xml-parser`
+        const xmlProblems = this.validateXML(document);
+        // Validate the document using the JSON Schema
+        const jsonSchemaProblems = this.validateJSONSchema(document);
+        // Return the problems
+        return [...xmlProblems, ...jsonSchemaProblems];
     }
 
     /**
@@ -54,32 +78,54 @@ export abstract class XMLDiagnostics extends Diagnostics {
     }
 
     /**
-     * Updates the {@link diagnostics| diagnostics collection} for the given document
-     * 
-     * If diagnostics for the given document are {@link allowDiagnostics | allowed},
-     * then the diagnostics collection is updated with the {@link createProblems | problems}. Otherwise,
-     * the diagnostics collection is cleared. 
-     * 
-     * @param document The document to update diagnostics for. (see {@linkcode vscode.TextDocument})
+     * Validates the XML document against the JSON Schema and returns an array of diagnostics for any errors found.
+     * @param document The TextDocument to validate.
+     * @returns An array of vscode.Diagnostic objects representing the errors found in the document.
      */
-    protected override updateDiagnostics(document: vscode.TextDocument) {
-        if (document && this.allowDiagnostics(document)) { // If diagnostics are allowed for this document...
-            const diagnostics: vscode.Diagnostic[] = [];
+    private validateJSONSchema(document: vscode.TextDocument): vscode.Diagnostic[] {
+        const diagnostics: vscode.Diagnostic[] = [];
 
-            // Validate the document using the XML Validator from `fast-xml-parser`
-            const xmlValidationProblem = this.validateXML(document);
-            diagnostics.push(...xmlValidationProblem);
+        // Parse the XML document
+        const text = document.getText();
+        const parsedXML = new XMLParser({
+            ignoreAttributes: false,
+            attributeNamePrefix: "",
+            attributesGroupName: xml.attributesGroupName,
+            parseAttributeValue: true,
+        }).parse(text);
 
-            // Create the diagnostics for the document using the `createProblems` method
-            const problems = this.createProblems(document);
-            diagnostics.push(...problems);
+        // Validate the document using the JSON Schema
+        const results = new Validator().validate(parsedXML, this.schema);
 
-            // Set the diagnostics collection
-            this.diagnostics.set(document.uri, diagnostics);
+
+        // If the document is not valid, then we create a diagnostic for the error
+        if (!results.valid) {
+            results.errors.forEach(error => {
+
+                // Determine the line of the error from the path
+                const line = this.determineLineFromPath(document, error.path);
+
+                // Create a range for the error
+                const range = new vscode.Range(
+                    line,
+                    document.lineAt(line).firstNonWhitespaceCharacterIndex,
+                    line,
+                    Number.MAX_VALUE
+                );
+
+                // Create the diagnostic object and add it to the collection
+                const problem: vscode.Diagnostic = {
+                    range,
+                    message: error.message,
+                    severity: vscode.DiagnosticSeverity.Error,
+                    source: this.name,
+                };
+                diagnostics.push(problem);
+
+            });
         }
-        else { // ... otherwise, we clear the diagnostics collection
-            this.diagnostics.clear();
-        }
+
+        return diagnostics;
     }
 
 }
